ClosureWrapper: separate Args into new file
