+++ closure-compiler-wrapper/src/main/java/com/prezi/spaghetti/closure/ClosureWrapper.java
@@ -14,81 +14,20 @@ import com.google.javascript.jscomp.DiagnosticType;
 import com.google.javascript.jscomp.JSError;
 import com.google.javascript.jscomp.ModuleIdentifier;
 import com.google.javascript.jscomp.SourceFile;
-import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.io.Writer;
 import java.util.logging.Level;
 import java.util.ArrayList;
 import java.util.List;
-import org.kohsuke.args4j.CmdLineException;
-import org.kohsuke.args4j.CmdLineParser;
-import org.kohsuke.args4j.Option;
 
-class Args {
-    @Option(name="--js_output_file")
-    public File outputFile;
-
-    @Option(name="--create_source_map")
-    public File sourceMap = null;
-
-    @Option(name="--entry_point")
-    public List<String> entryPoints = new ArrayList<String>();
-
-    @Option(name="--js")
-    public List<String> inputPatterns = new ArrayList<String>();
-
-    @Option(name="--externs")
-    public List<String> externsPatterns = new ArrayList<String>();
-
-    @Option(name="--compilation_level")
-    public String compilationLevel = "SIMPLE";
-
-    @Option(name="--concat")
-    public boolean concat = false;
-
-    @Option(name="--es5")
-    public boolean es5 = false;
-}
 
 class ClosureWrapper {
 
     static DiagnosticType EARLY_REFERENCE = findDiagnosticType("JSC_REFERENCE_BEFORE_DECLARE");
 
-    // VariableReferenceCheck is a protected class, so we have to access
-    // VariableReferenceCheck.EARLY_REFERENCE the hacky way.
-    static DiagnosticType findDiagnosticType(String key) {
-        for (DiagnosticType t : DiagnosticGroups.CHECK_VARIABLES.getTypes()) {
-            if (key.equals(t.key)) {
-                return t;
-            }
-        }
-
-        throw new RuntimeException("Cannot locate EARLY_REFERENCE");
-    }
-
-
-    private static List<ModuleIdentifier> getEntryPoints(List<String> entryFiles) {
-        List<ModuleIdentifier> entryPoints = new ArrayList<ModuleIdentifier>();
-        for (String s : entryFiles) {
-            entryPoints.add(ModuleIdentifier.forFile(s));
-        }
-        return entryPoints;
-    }
-
     public static void main(String[] args) throws IOException {
-        Args parsedArgs = new Args();
-        CmdLineParser parser = new CmdLineParser(parsedArgs);
-
-        try {
-            parser.parseArgument(args);
-        } catch (CmdLineException e) {
-            System.err.println(e.getMessage());
-            parser.printUsage(System.err);
-            return;
-        }
-
-        runCompiler(parsedArgs);
+        runCompiler(Args.parse(args));
     }
 
     private static void runCompiler(Args args) throws IOException {
@@ -182,5 +121,26 @@ class ClosureWrapper {
                 System.out.println("Wrote: " + args.sourceMap.getAbsolutePath());
             }
         }
+    }
+
+    // VariableReferenceCheck is a protected class, so we have to access
+    // VariableReferenceCheck.EARLY_REFERENCE the hacky way.
+    static DiagnosticType findDiagnosticType(String key) {
+        for (DiagnosticType t : DiagnosticGroups.CHECK_VARIABLES.getTypes()) {
+            if (key.equals(t.key)) {
+                return t;
+            }
+        }
+
+        throw new RuntimeException("Cannot locate EARLY_REFERENCE");
+    }
+
+
+    private static List<ModuleIdentifier> getEntryPoints(List<String> entryFiles) {
+        List<ModuleIdentifier> entryPoints = new ArrayList<ModuleIdentifier>();
+        for (String s : entryFiles) {
+            entryPoints.add(ModuleIdentifier.forFile(s));
+        }
+        return entryPoints;
     }
 }
